+++ app/external-api/src/main/java/cos/peerna/controller/notification/response/NotificationResponse.java
@@ -1,13 +1,13 @@
 package cos.peerna.controller.notification.response;
 
-import cos.peerna.domain.user.dto.data.UserProfileData;
+import cos.peerna.domain.user.dto.UserProfile;
 import java.time.LocalDate;
 
 public record NotificationResponse(
 		Long notificationId,
 		String type,
 		String answer,
-		UserProfileData sender,
+		UserProfile sender,
 		String msg,
 		LocalDate time
 ) {}
