+++ domain/integrated/src/main/java/cos/peerna/domain/notification/dto/data/NotificationData.java
@@ -1,6 +1,6 @@
 package cos.peerna.domain.notification.dto.data;
 
-import cos.peerna.domain.user.dto.data.UserProfileData;
+import cos.peerna.domain.user.dto.UserProfile;
 import lombok.Builder;
 import lombok.Data;
 
@@ -12,7 +12,7 @@ public class NotificationData {
 	private Long        notificationId;
 	private String		type;
 	private String      answer;
-	private UserProfileData sender;
+	private UserProfile sender;
 	private String      msg;
 	private LocalDate   time;
 }
