+++ domain/integrated/src/main/java/cos/peerna/domain/room/dto/RoomResponseDto.java
@@ -1,6 +1,6 @@
 package cos.peerna.domain.room.dto;
 
-import cos.peerna.domain.user.dto.data.UserProfileData;
+import cos.peerna.domain.user.dto.UserProfile;
 import cos.peerna.domain.problem.model.Problem;
 import lombok.Builder;
 import lombok.Data;
@@ -11,6 +11,5 @@ public class RoomResponseDto {
     private Integer roomId;
     private Long    historyId;
     private Problem problem;
-    private UserProfileData peer;
-
+    private UserProfile peer;
 }
